set up chain for sinatra
