+++ test/multiverse/suites/sinatra/Envfile
@@ -1,3 +1,5 @@
+instrumentation_methods :chain#, :prepend
+
 rack_test_version = RUBY_VERSION < "2.2.2" ? "< 0.8.0" : ">= 0.8.0"
 
 if RUBY_VERSION >= '2.2.0'
